+++ CHANGELOG.md
@@ -19,11 +19,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Security
 
-## [0.2.1] - 2025-03-24
+## [0.2.1] - 2025-04-16
 
-- Support file uploads.
+- Support user file uploads in message content.
 - Fix indefinite loading when same message is sent twice.
-- Add minimal styling to table.
+- Implement renderers to support varying assistant message content
 
 ## [0.2.0] - 2025-02-12